PJT/requirements.txt
Django>=4.0
mysqlclient
djangorestframework
gunicorn
python-dotenv
whitenoise
django-cors-headers
